refactor(registration): dedupe 3x688-line routes + zod → Effect Schema
The 2024/2025/2026 registration form routes were byte-identical 688-line
files differing only by the heading year. Extract one shared
`RegistrationForm` component parameterized by `year` (a plain module under
`app/routes/($lang)+/`, not referenced in routes.ts, so it is not itself a
route). The three routes become 52-line thin wrappers: `meta`, a
`routeHandler` loader (locale via `ReactRouterContext` params +
`getCurrentConference`, which stays a pure lookup), a typed `routeAction`
no-op action (submission handling is a deliberate, pending product
decision), and `<RegistrationForm year={…} />`. Net ~1,080 fewer lines.

Port the zod `RegistrationSchema` to an Effect Schema **form-data codec**
(`registration-schema.ts`): strings stay strings; a custom
`StringToBoolean` codec decodes the tokens the form actually submits
(radio/checkbox `"true"`/`"false"` for meals/firstTimeAttending/tos;
attribute-less checkbox `"on"` for the volunteer flags — there is no
`Schema.BooleanFromString` in effect beta.60); literal unions for
gender/outreach/merch; an optional nested `parent` struct; optional
volunteer flags via `optionalKey`. Array fields use `Schema.mutable` so
conform's `getFieldList`/`FieldName` accept them. Client validation runs
through `Schema.toStandardSchemaV1`; the `coerceFormValue` interim and all
zod / `@conform-to/zod` imports are dropped from these routes.

Accepted drift (per plan): registration validation messages move off zod
defaults onto translation keys (`registration.form.email.error`,
`registration.form.website.required`) already present in en+fr — the prior
zod boolean coercion only accepted `"on"`, so the `"true"`/`"false"`
radio/checkbox booleans were effectively unvalidated; the new codec
validates them correctly on this client-only path.

Tests: `registration-schema.test.ts` covers the boolean coercion, the
discriminated attendee/exhibitor union, email/url validation, and the
nested-array conform field-name path (`registrants[0].email`, never the
dotted form). Gate green (typecheck/lint/test/build). Dev-server smoke
(PORT=3942): GET /{2024,2025,2026}/form and /fr/2026/form all 200 with the
correct per-year heading; POST to the no-op action returns 200; the
`registrants[0].*` field names and Add Registrant intent render.

Codex-review: no blocking issues, no nits — confirmed the diff matches the
C5 plan shape (thin wrappers, shared component, Effect Schema codec via
toStandardSchemaV1, no-op actions preserved, boolean tokens true/false/on
covered) and the 4 schema tests pass.
